+++ tasks/__init__.py
@@ -3,3 +3,4 @@ from .fetch_nasa_fires import fetch_nasa_fires
 from .fetch_openaq_latest import fetch_openaq_latest  
 from .fetch_noaa_ocean import fetch_noaa_ocean_data
 from .fetch_weather import fetch_weather_data
+from .fetch_biodiversity import fetch_biodiversity_data